Moved some code into database/core.rs.
